kernel.spec: Support to build perf RMP for python3 only system
kernel tools, such as perf, depend on python.
Although python3 is the choice in latest openEuler innovations, python2
is the default version in 20-03 LTS series. It makes sense to build perf
for systems with python2 or python3.

The original kernel spec can't work to generate perf for python3.
This patch makes some changes on spec to fix. A new build option of 'python3'
is introduced. The default is only to build python2 related perf to keep
20-03 compatibility. If a python3 related perf RPM is needed, please run
rpmbuild with '--with python3'. This means one more kernel RPM building
should be performed.
It will take extra time for the 2nd full kernel building.

Probably the better way is to break the spec into two spec files, one is
specific for kernel tools including perf, another is only for kernel
core components. It can save building time on the second kernel RPM
buidling. But this means a new repo for kernel tools should be
created with more effort. If python2 dependent perf RPM is needed for
long time, then we can make the separation.

Signed-off-by: Zhichang Yuan <erik.yuan@arm.com>