riscv: Add riscv-kernel patch
riscv-kernel[1] is to add support for multiple RISC-V SoCs on the basis of
OLK-6.6 and enrich the openEuler riscv ecosystem.

Currently, riscv-kernel has added support for sg2042[2] and th1520.
After applying this patch, we can build the same kernel to start openEuler
on qemu, sg2042, and th1520 platforms.

Build and boot testing passed on qemu, Milk-V Pioneer[3] and Lpi4a[4].

Link: https://gitee.com/openeuler/riscv-kernel/tree/OLK-6.6 [1]
Link: https://www.sophgo.com/product/introduce/sg2042.html [2]
Link: https://milkv.io/zh/pioneer [3]
Link: https://sipeed.com/licheepi4a [4]
Signed-off-by: Mingzheng Xing <xingmingzheng@iscas.ac.cn>